The remaining protection groups have been updated. The plugin now has exactly 49 active flags. Added flags: • command-use — general command usage. • fishing — fishing and casting across the plot boundary. • elytra — initiating elytra flight. • flight — initiating regular flight. • special-weapons — tridents and maces. • weather — lightning strikes and lightning damage. • bed-use — using beds. • crafting — crafting table and crafter. • enchanting — enchanting table. • respawn-anchor — using the respawn anchor. • conduit-effects — effects applied by the conduit.

+++ src/main/kotlin/pl/syntaxdevteam/plotsx/hooks/CoreProtectHook.kt
@@ -76,7 +76,7 @@ class CoreProtectHook(private val plugin: PlotsX) {
         val api = coreProtectAPI ?: return
         val seconds = timeSeconds.toInt()
         val world = Bukkit.getWorld(plot.world) ?: return
-        val centerY = world.spawnLocation.y.toDouble()
+        val centerY = world.spawnLocation.y
         val center = Location(world, plot.x.toDouble(), centerY, plot.z.toDouble())
 
         Bukkit.getScheduler().runTaskAsynchronously(plugin, Runnable {
@@ -100,7 +100,7 @@ class CoreProtectHook(private val plugin: PlotsX) {
         val api = coreProtectAPI ?: return emptyList()
         val seconds = timeSeconds.toInt()
         val world = Bukkit.getWorld(plot.world) ?: return emptyList()
-        val centerY = world.spawnLocation.y.toDouble()
+        val centerY = world.spawnLocation.y
         val center = Location(world, plot.x.toDouble(), centerY, plot.z.toDouble())
 
         return api.performLookup(
@@ -123,7 +123,7 @@ class CoreProtectHook(private val plugin: PlotsX) {
         val api = coreProtectAPI ?: return
         val seconds = timeSeconds.toInt()
         val world = Bukkit.getWorld(plot.world) ?: return
-        val centerY = world.spawnLocation.y.toDouble()
+        val centerY = world.spawnLocation.y
         val center = Location(world, plot.x.toDouble(), centerY, plot.z.toDouble())
 
         Bukkit.getScheduler().runTaskAsynchronously(plugin, Runnable {
